backend/requirements.txt
django==3.2.6
djangorestframework==3.12.4
drf-yasg2==1.19.4
django-filter==21.1
python-dotenv==0.19.1
psycopg2-binary==2.9.1
djangorestframework-simplejwt==4.6.0
Pillow==8.4.0
boto3==1.20.3
django-storages==1.12.3

pytest==6.2.3
pytest-django==4.2.0
pylint==2.7.4